bench/requirements.txt
pdfplumber==0.9.0
pdfminer.six==20221105
camelot-py[cv]==0.11.0
tabula-py==2.9.3
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.4
PyPDF2==3.0.1